Convert sticky post to grid / fix footer indent and h2
Sticky is back to being a grid as it cleans up the code.

Use :not() for solo and footer indent removal.

Add `h2` section title outside and inside `section` until Firefox fixes support for column-span. A little dirty but I can live with it. Delete from articles and CSS when the fix lands.